Fix a bug and accept abs operation.

+++ src/main/java/ua/in/boyaryn/grammar/expression/Expression.java
@@ -1,5 +1,7 @@
 package ua.in.boyaryn.grammar.expression;
 
+import java.math.BigDecimal;
+
 public interface Expression {
-    double evaluate();
+    BigDecimal evaluate();
 }
