+++ src/main/java/ua/in/boyaryn/grammar/expression/UnOpExpression.java
@@ -2,16 +2,18 @@ package ua.in.boyaryn.grammar.expression;
 
 import ua.in.boyaryn.grammar.operator.UnaryOperator;
 
+import java.math.BigDecimal;
+
 public class UnOpExpression implements Expression {
 
-    private final double value;
+    private final BigDecimal value;
 
     public UnOpExpression(Expression operand, UnaryOperator operator) {
         this.value = operator.action(operand.evaluate());
     }
 
     @Override
-    public double evaluate() {
+    public BigDecimal evaluate() {
         return value;
     }
 }
